Flash messages for roles and collaborators

+++ src/main/java/com/techdegree/instateam/web/controller/RoleController.java
@@ -1,7 +1,9 @@
 package com.techdegree.instateam.web.controller;
 
+import com.techdegree.instateam.exception.NotFoundException;
 import com.techdegree.instateam.model.Role;
 import com.techdegree.instateam.service.RoleService;
+import com.techdegree.instateam.web.FlashMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -42,7 +44,6 @@ public class RoleController {
                              BindingResult result,
                              RedirectAttributes redirectAttributes) {
         if (result.hasErrors()) {
-            System.out.println(result.hasErrors());
             // we add flash to remember user's input
             redirectAttributes.addFlashAttribute(
                     "org.springframework.validation.BindingResult.newRole",
@@ -53,6 +54,11 @@ public class RoleController {
         }
         // save role to database
         roleService.save(role);
+        // success flash message
+        redirectAttributes.addFlashAttribute("flash",
+                new FlashMessage("New Role: '" + role.getName() +
+                        "' was successfully added!",
+                        FlashMessage.Status.SUCCESS));
         // redirect back to same page
         return "redirect:/roles";
     }
@@ -64,6 +70,10 @@ public class RoleController {
         // here we check if user made wrong input, and show his wrong data
         if (!model.containsAttribute("role")) {
             Role role = roleService.findById(roleId);
+            // if role is not found we throw exception showing error page
+            if (role == null) {
+                throw new NotFoundException("Role not found");
+            }
             model.addAttribute("role", role);
         }
         return "role/role-details";
@@ -85,17 +95,40 @@ public class RoleController {
             // redirect to detail role page
             return "redirect:/roles/" + roleId;
         }
+        // get old role from database
+        String oldRoleName = roleService.findById(roleId).getName();
         // save role to database
         roleService.save(role);
+        redirectAttributes.addFlashAttribute("flash",
+                new FlashMessage(
+                        "Role: '" + oldRoleName +
+                        "' was successfully updated to '"
+                        + role.getName() +
+                        "'",
+                        FlashMessage.Status.SUCCESS));
         // redirect to page with all roles
         return "redirect:/roles";
     }
 
     //Delete role
     @RequestMapping(value = "/roles/{roleId}/delete")
-    public String deleteRole(@PathVariable int roleId) {
+    public String deleteRole(
+            @PathVariable int roleId,
+            RedirectAttributes redirectAttributes) {
+        // try to find role by id
         Role role = roleService.findById(roleId);
+        // if not found throw not found exception
+        if (role == null) {
+            throw new NotFoundException("Role not Found");
+        }
+        // delete from database
         roleService.delete(role);
+        // success flash message
+        redirectAttributes.addFlashAttribute("flash",
+                new FlashMessage("Role: '" + role.getName() +
+                        "' was successfully deleted!",
+                        FlashMessage.Status.SUCCESS));
+        // redirect back to page with roles
         return "redirect:/roles";
     }
 }
